+++ src/main.cpp
@@ -14,6 +14,8 @@ class $modify(LevelEditorLayer) {
 
         Variables::updateSettings();
 
+        GameManager::sharedState()->setGameVariable("0058", false); // hide default 
+
         auto layer = CCLayer::create();
         layer->setPosition(0.0f, 0.0f);
         layer->setZOrder(-1500);