+++ changelog.md
@@ -1,2 +1,10 @@
+## v1.1.0
+- **Added** layers are faded now
+- **Added** a lil bit of code cleanup
+- **Added** automatically disabling default duration indicators
+
+- **Fixed** culling
+- **Fixed** pulse and sfx indicator draw order
+
 ## v1.0.0
 - Initial Release